gnl now with barebones libft and no macros (less liabilities)

+++ libft/srcs/ft_count_digits.c
@@ -6,7 +6,7 @@
 /*   By: olkovale <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/06/18 20:17:46 by olkovale          #+#    #+#             */
-/*   Updated: 2017/07/14 12:11:24 by olkovale         ###   ########.fr       */
+/*   Updated: 2017/06/19 21:50:01 by olkovale         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 #include "libft.h"
 
-size_t		ft_count_digits(long value)
+size_t		ft_count_digits(int value)
 {
 	size_t	count;
